Fix data config with addons
